minors

+++ src/laser/repump.rs
@@ -37,7 +37,7 @@ impl<'a> System<'a> for RepumpSystem {
             None => (),
             Some(repump) => {
                 for (ent, num) in (&ent, &num).join() {
-                    if repump.if_loss(num.value) {
+                    if repump.if_loss(num.value as f64) {
                         lazy.insert(ent, Dark {})
                     }
                 }
